Reorganize imports, since model folders structure changed
